refactor(web-scraping): move cluster extraction strategy
