+++ tldw_Server_API/app/core/Web_Scraping/extraction/strategies/__init__.py
@@ -1,6 +1,7 @@
 """Canonical extraction strategy implementations."""
 
+from .cluster import extract_cluster_entities
 from .jsonld import extract_jsonld_entities
 from .regex import extract_regex_entities
 
-__all__ = ["extract_jsonld_entities", "extract_regex_entities"]
+__all__ = ["extract_cluster_entities", "extract_jsonld_entities", "extract_regex_entities"]
